setup enhancements WIP
