+++ autosportlabs/racecapture/views/tracks/tracksview.py
@@ -25,6 +25,7 @@ kivy.require('1.10.0')
 from kivy.properties import NumericProperty, ObjectProperty, BooleanProperty, ListProperty
 from kivy.uix.boxlayout import BoxLayout
 from kivy.uix.stacklayout import StackLayout
+from kivy.uix.anchorlayout import AnchorLayout
 from kivy.clock import Clock
 from kivy.uix.label import Label
 from kivy.uix.popup import Popup
@@ -45,6 +46,7 @@ from autosportlabs.racecapture.geo.geopoint import GeoPoint
 from iconbutton import LabelIconButton
 from autosportlabs.widgets.scrollcontainer import ScrollContainer
 from autosportlabs.racecapture.tracks.trackmanager import TrackManager, TrackMap
+from autosportlabs.racecapture.config.rcpconfig import Track
 
 Builder.load_file('autosportlabs/racecapture/views/tracks/tracksview.kv')
 
@@ -345,7 +347,7 @@ class TracksBrowser(BoxLayout):
         if self.tracksGrid:
             for trackView in self.tracksGrid.children:
                 trackView.setSelected(value)
-                
+
 class TrackSelectView(StackLayout):
 
     Builder.load_string("""
@@ -373,10 +375,10 @@ class TrackSelectView(StackLayout):
         self._track_manager = track_manager
         self.ids.track_browser.current_location = current_location
         self.ids.track_browser.init_view()
-        
+
     def on_track_selected(self, track):
         pass
-    
+
     def track_selected(self, instance, tracks):
         if len(tracks) > 0:
             # Tracks are a set, we only want 1 but we don't want to modify the original
@@ -387,4 +389,201 @@ class TrackSelectView(StackLayout):
             self.dispatch('on_track_selected', track)
         else:
             self.selected_track = None
-                
+
+class EmptyTrackDbView(BoxLayout):
+    Builder.load_string("""
+<EmptyTrackDbView>:
+    BoxLayout:
+        orientation: 'vertical'
+        pos_hint: {'center_x': .5, 'center_y': .5}        
+        Label:
+            font_size: dp(20)
+            text: 'No tracks selected'
+        Label:
+            height: dp(30)
+        Label:
+            text: 'Press the (+) button to select your favorite tracks'    
+    """)
+    def __init__(self, **kwargs):
+        super(EmptyTrackDbView, self).__init__(**kwargs)
+
+class TrackSelectionPopup(BoxLayout):
+    Builder.load_string("""
+<TrackSelectionPopup>:
+    orientation: 'vertical'
+    TracksBrowser:
+        trackHeight: dp(200)
+        id: browser
+        size_hint_y: 0.90
+    BoxLayout:
+        size_hint_y: 0.1
+        orientation: 'horizontal'
+        IconButton:
+            text: '\357\200\214'
+            color: ColorScheme.get_accent()
+            on_release: root.confirm_add_tracks()    
+    """)
+    def __init__(self, **kwargs):
+        super(TrackSelectionPopup, self).__init__(**kwargs)
+        self.register_event_type('on_tracks_selected')
+        track_manager = kwargs.get('track_manager', None)
+        track_browser = self.ids.browser
+        track_browser.set_trackmanager(track_manager)
+        track_browser.init_view()
+        self.track_browser = track_browser
+
+    def on_tracks_selected(self, selectedTrackIds):
+        pass
+
+    def confirm_add_tracks(self):
+        self.dispatch('on_tracks_selected', self.track_browser.selectedTrackIds)
+
+TRACK_DB_ITEM_VIEW_KV = """
+<TrackDbItemView>:
+    BoxLayout:
+        orientation: 'horizontal'
+        TrackInfoView:
+            size_hint_x: 0.92
+            id: trackinfo
+        AnchorLayout:
+            size_hint_x: 0.08
+            IconButton:
+                size_hint: (0.5, 0.15)
+                text: '\357\200\224'
+                on_release: root.remove_track()
+"""
+
+class TrackDbItemView(BoxLayout):
+    Builder.load_string(TRACK_DB_ITEM_VIEW_KV)
+    def __init__(self, **kwargs):
+        super(TrackDbItemView, self).__init__(**kwargs)
+        track = kwargs.get('track', None)
+        self.index = kwargs.get('index', 0)
+        self.ids.trackinfo.setTrack(track)
+        self.track = track
+        self.register_event_type('on_remove_track')
+
+    def on_remove_track(self, index):
+        pass
+
+    def remove_track(self):
+        self.dispatch('on_remove_track', self.index)
+
+class TrackCollectionScreen(AnchorLayout):
+    Builder.load_string("""
+<TrackCollectionScreen>:
+    AnchorLayout:
+        spacing: dp(10)
+        ScrollContainer:
+            canvas.before:
+                Color:
+                    rgba: 0.05, 0.05, 0.05, 1
+                Rectangle:
+                    pos: self.pos
+                    size: self.size
+            size_hint_y: 0.95
+            id: scrltracks
+            do_scroll_x:False
+            do_scroll_y:True
+            GridLayout:
+                id: tracksgrid
+                padding: [dp(10), dp(10)]
+                spacing: [dp(10), dp(10)]
+                size_hint_y: None
+                cols: 1
+    AnchorLayout:
+        anchor_y: 'bottom'
+        IconButton:
+            color: ColorScheme.get_accent()
+            size_hint: (None, None)
+            height: root.height * .15
+            text: u'\uf055'
+            on_release: root.on_add_track_db()
+            disabled: True
+            id: addtrack    
+    """)
+    track_manager = ObjectProperty(None)
+    TRACK_ITEM_MIN_HEIGHT = 200
+
+    def __init__(self, **kwargs):
+        super(TrackCollectionScreen, self).__init__(**kwargs)
+        self._track_db = None
+        self.register_event_type('on_modified')
+
+    def on_modified(self, *args):
+        pass
+
+    def on_config_updated(self, track_db):
+        self._track_db = track_db
+        self.init_tracks_list()
+
+    def on_track_manager(self, instance, value):
+        self.track_manager = value
+        self.init_tracks_list()
+
+    def on_add_track_db(self):
+
+        def on_tracks_selected(instance, selected_track_ids):
+            if self._track_db:
+                failures = False
+                for trackId in selected_track_ids:
+                    trackMap = self.track_manager.get_track_by_id(trackId)
+                    if trackMap:
+                        startFinish = trackMap.start_finish_point
+                        if startFinish and startFinish.latitude and startFinish.longitude:
+                            Logger.info("TrackConfigView:  adding track " + str(trackMap))
+                            track = Track.fromTrackMap(trackMap)
+                            self._track_db.tracks.append(track)
+                        else:
+                            failures = True
+                if failures:
+                    alertPopup('Cannot Add Tracks', 'One or more tracks could not be added due to missing start/finish points.\n\nPlease check for track map updates and try again.')
+                self.init_tracks_list()
+                popup.dismiss()
+                self._track_db.stale = True
+                self.dispatch('on_modified')
+
+        content = TrackSelectionPopup(track_manager=self.track_manager)
+        popup = Popup(title='Add Race Tracks', content=content, size_hint=(0.9, 0.9))
+        content.bind(on_tracks_selected=on_tracks_selected)
+        popup.open()
+
+    def init_tracks_list(self):
+        if self.track_manager and self._track_db:
+            matched_tracks = []
+            for track in self._track_db.tracks:
+                matched_track = self.track_manager.find_track_by_short_id(track.trackId)
+                if matched_track:
+                    matched_tracks.append(matched_track)
+
+            grid = self.ids.tracksgrid
+            grid.clear_widgets()
+            if len(matched_tracks) == 0:
+                grid.add_widget(EmptyTrackDbView())
+                grid.height = dp(self.TRACK_ITEM_MIN_HEIGHT)
+            else:
+                grid.height = dp(self.TRACK_ITEM_MIN_HEIGHT) * (len(matched_tracks) + 1)
+                index = 0
+                for track in matched_tracks:
+                    track_db_view = TrackDbItemView(track=track, index=index)
+                    track_db_view.bind(on_remove_track=self.on_remove_track)
+                    track_db_view.size_hint_y = None
+                    track_db_view.height = dp(self.TRACK_ITEM_MIN_HEIGHT)
+                    grid.add_widget(track_db_view)
+                    index += 1
+
+            self.disableView(False)
+
+    def on_remove_track(self, instance, index):
+            try:
+                del self._track_db.tracks[index]
+                self.init_tracks_list()
+                self._track_db.stale = True
+                self.dispatch('on_modified')
+
+            except Exception as detail:
+                Logger.error('AutomaticTrackConfigScreen: Error removing track from list ' + str(detail))
+
+    def disableView(self, disabled):
+        self.ids.addtrack.disabled = disabled
+
